# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A secure, microservice-style Ethereum wallet API in Go: wallet creation, transaction signing, and balance checking on the Sepolia testnet.\nPrivate keys are stored encrypted (Web3 Secret Storage). Endpoints are protected by JWT bearer tokens (OAuth2 client-credentials) with role-based access control.",
        "title": "Web3 Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/token": {
            "post": {
                "description": "Exchanges client_id/client_secret for a short-lived JWT bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Issue an access token (OAuth2 client credentials)",
                "parameters": [
                    {
                        "description": "Client credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.tokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ameerhamza_web3-wallet-api_internal_auth.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Always returns 200 if the process is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.healthResponse"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Returns 200 when the service can serve traffic. RPC connectivity is reported but not required (offline features remain available).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.healthResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Builds, signs, and broadcasts an ETH transfer from a managed wallet. Requires RPC connectivity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Send a value transfer",
                "parameters": [
                    {
                        "description": "Transfer parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.sendRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_ameerhamza_web3-wallet-api_internal_transaction.Receipt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/wallets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all accounts held by the service. Requires the admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "List managed wallets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_ameerhamza_web3-wallet-api_internal_wallet.Wallet"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new Ethereum account; the private key is stored encrypted (Web3 Secret Storage) and never returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Create a new wallet",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ameerhamza_web3-wallet-api_internal_wallet.Wallet"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/wallets/{address}/balance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reads the latest on-chain balance for any address. Requires RPC connectivity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Get an address balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethereum address (0x...)",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.balanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_ameerhamza_web3-wallet-api_internal_auth.Token": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "github_com_ameerhamza_web3-wallet-api_internal_transaction.Receipt": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "gas_limit": {
                    "type": "integer"
                },
                "gas_price_wei": {
                    "type": "string"
                },
                "nonce": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                },
                "value_wei": {
                    "type": "string"
                }
            }
        },
        "github_com_ameerhamza_web3-wallet-api_internal_wallet.Wallet": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.balanceResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x71C7656EC7ab88b098defB751B7401B5f6d8976F"
                },
                "balance_eth": {
                    "type": "string",
                    "example": "1.000000000000000000"
                },
                "balance_wei": {
                    "type": "string",
                    "example": "1000000000000000000"
                }
            }
        },
        "internal_api_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid address"
                }
            }
        },
        "internal_api_handlers.healthResponse": {
            "type": "object",
            "properties": {
                "rpc_online": {
                    "type": "boolean",
                    "example": true
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "internal_api_handlers.sendRequest": {
            "type": "object",
            "required": [
                "from",
                "to",
                "value_wei"
            ],
            "properties": {
                "from": {
                    "type": "string",
                    "example": "0x71C7656EC7ab88b098defB751B7401B5f6d8976F"
                },
                "gas_limit": {
                    "type": "integer",
                    "example": 21000
                },
                "to": {
                    "type": "string",
                    "example": "0x71C7656EC7ab88b098defB751B7401B5f6d8976F"
                },
                "value_wei": {
                    "type": "string",
                    "example": "1000000000000000"
                }
            }
        },
        "internal_api_handlers.tokenRequest": {
            "type": "object",
            "required": [
                "client_id",
                "client_secret"
            ],
            "properties": {
                "client_id": {
                    "type": "string",
                    "example": "demo-client"
                },
                "client_secret": {
                    "type": "string",
                    "example": "dev-insecure-client-secret-change-me"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}